+++ lib/kpabase/SettingsData.h
@@ -1,5 +1,6 @@
 // SPDX-FileCopyrightText: 2003-2020 The KPhotoAlbum Development Team
 // SPDX-FileCopyrightText: 2021 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+// SPDX-FileCopyrightText: 2022 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
 //
 // SPDX-License-Identifier: GPL-2.0-or-later
 
@@ -258,6 +259,8 @@ signals:
     void actualThumbnailSizeChanged(int);
     void histogramScaleChanged();
     void colorSchemeChanged();
+    void displayLabelsChanged(bool);
+    void displayCategoriesChanged(bool);
 
 private:
     SettingsData(const QString &imageDirectory, DB::UIDelegate &delegate);